requirements.txt
Flask==3.1.0
mercadopago==2.2.3
pycryptodome==3.21.0
python-dotenv==1.1.0
